+++ hashmap.h
@@ -2,10 +2,9 @@
 #define HASHMAP_H_
 
 #include <stdbool.h>
-#include <stdlib.h>
 #include <stdint.h>
 
-typedef uint64_t hash_t;
+typedef uint32_t hash_t;
 
 struct hash_entry {
 	struct hash_entry *next;
@@ -14,8 +13,8 @@ struct hash_entry {
 
 struct hash_table {
 	/** Must be 2^n - 1 */
-	size_t mask;
-	size_t size;
+	uint32_t mask;
+	uint32_t size;
 	struct hash_entry **data;
 };
 
@@ -25,7 +24,7 @@ struct hash_map {
 
 typedef bool (*hash_entry_cmp_fn)(const void *, const void *);
 
-void hash_map_init(struct hash_map *m, size_t cap);
+void hash_map_init(struct hash_map *m, uint32_t cap);
 
 struct hash_entry *hash_map_get(
 	const struct hash_map *m,
